"Added FootNote component to several pages and updated Table component to accept FootNote props"
This commit adds the FootNote component to several pages, including AnalyticsBody, ExplorePage, ReservationsPage, and ReviewsPage. It also updates the Table component to accept FootNote props, which includes a list of notes to be displayed in the FootNote component. The FootNote component is used to provide additional information or context to the user.